backend/test_requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-flask>=1.2.0
pytest-xdist>=3.3.1
coverage>=7.2.0
faker>=19.3.0
factory-boy>=3.3.0